+++ feishukit/feishu_im/components/__init__.py
@@ -1,0 +1,4 @@
+from ._group import GroupMixin
+from ._message import MessageMixin
+
+__all__ = ["GroupMixin", "MessageMixin"]